+++ packages/horizon-experience/src/horizon_experience/exceptions/__init__.py
@@ -1,0 +1,5 @@
+"""Experience exceptions."""
+
+from horizon_experience.exceptions.input import UserCancelled
+
+__all__ = ["UserCancelled"]