+++ packages/horizon-experience/src/horizon_experience/menus/__init__.py
@@ -1,0 +1,5 @@
+"""Experience menus."""
+
+from horizon_experience.menus.lab import print_menu
+
+__all__ = ["print_menu"]